Redesigned Network tab to not be a giant mess.
