+++ src/views/view.hpp
@@ -40,6 +40,7 @@ namespace big
 		static void self();
 		static void animations();
 		static void network();
+		static void network_controls();
 		static void chat();
 		static void missions();
 		static void player_database();
